fix project file and some source code to avoid compiler error

+++ fix_unistd.h
@@ -103,7 +103,7 @@ __forceinline int UNISTD_stime(const time_t *tp ){
 	SYSTEMTIME st;
     LONGLONG ll = Int32x32To64(*tp, 10000000) + 116444736000000000;
     ft.dwLowDateTime = (DWORD) ll;
-    ft.dwHighDateTime = (DWORD)ll >>32;
+    ft.dwHighDateTime = (unsigned __int64)ll >>32;
     FileTimeToSystemTime(&ft,&st);
 	return SetSystemTime(&st);
 }
